pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-radar-dolma-wrapper"
version = "0.1.0"
description = "Download datasets, normalize/sanitize with Dolma toolkit triggers, and upload to S3 for LLM training."
requires-python = ">=3.10"
dependencies = [
  "datasets>=2.16.0",
  "boto3",
  "pyyaml",
]

[project.scripts]
drdw = "data_radar_dolma_wrapper.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
